LayoutTests/http/tests/fetch/fetch-redirect-same-origin-authorization-expected.txt

Tests that the Authorization header is present on same-origin redirect requests.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS authorization is "Foo Bar"
PASS successfullyParsed is true

TEST COMPLETE
